requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
pillow>=11.2.1
pillow-heif==1.0.0
piexif>=1.1.3
exifread>=3.3.1
geopy==2.3.0
pytz>=2023.3
ffmpeg-python>=0.2.0
pywin32; platform_system == "Windows"